--- a/storage/reads/table.go
+++ b/storage/reads/table.go
@@ -31,6 +31,7 @@ type table struct {
 	err error
 
 	cancelled, used int32
+	cache           *tagsCache
 	alloc           *memory.Allocator
 }
 
@@ -40,6 +41,7 @@ func newTable(
 	key flux.GroupKey,
 	cols []flux.ColMeta,
 	defs [][]byte,
+	cache *tagsCache,
 	alloc *memory.Allocator,
 ) table {
 	return table{
@@ -49,6 +51,7 @@ func newTable(
 		tags:   make([][]byte, len(cols)),
 		defs:   defs,
 		cols:   cols,
+		cache:  cache,
 		alloc:  alloc,
 	}
 }
@@ -197,30 +200,15 @@ func (t *table) appendTags(cr *colReader) {
 	for j := range t.cols {
 		v := t.tags[j]
 		if v != nil {
-			b := arrow.NewStringBuilder(t.alloc)
-			b.Reserve(cr.l)
-			b.ReserveData(cr.l * len(v))
-			for i := 0; i < cr.l; i++ {
-				b.Append(v)
-			}
-			cr.cols[j] = b.NewArray()
-			b.Release()
+			cr.cols[j] = t.cache.GetTag(string(v), cr.l, t.alloc)
 		}
 	}
 }
 
 // appendBounds fills the colBufs for the time bounds
 func (t *table) appendBounds(cr *colReader) {
-	bounds := []execute.Time{t.bounds.Start, t.bounds.Stop}
-	for j := range []int{startColIdx, stopColIdx} {
-		b := arrow.NewIntBuilder(t.alloc)
-		b.Reserve(cr.l)
-		for i := 0; i < cr.l; i++ {
-			b.UnsafeAppend(int64(bounds[j]))
-		}
-		cr.cols[j] = b.NewArray()
-		b.Release()
-	}
+	start, stop := t.cache.GetBounds(t.bounds, cr.l, t.alloc)
+	cr.cols[startColIdx], cr.cols[stopColIdx] = start, stop
 }
 
 func (t *table) closeDone() {