@jsternberg 2019-11-25
Performance metrics from the benchmark.

### Only cache start/stop columns

```
jsternberg:[~/g/s/g/i/influxdb] $ (perf/table-read $<>) benchcmp old.txt new.txt
benchmark                  old ns/op      new ns/op      delta
BenchmarkReadFilter-12     1291650820     1135894099     -12.06%

benchmark                  old allocs     new allocs     delta
BenchmarkReadFilter-12     1350688        1170693        -13.33%

benchmark                  old bytes      new bytes     delta
BenchmarkReadFilter-12     1379473016     983649672     -28.69%
```

### Cache both tags and start/stop column

```
jsternberg:[~/g/s/g/i/influxdb] $ (perf/table-read $<>) benchcmp old.txt new2.txt
benchmark                  old ns/op      new ns/op     delta
BenchmarkReadFilter-12     1291650820     523629173     -59.46%

benchmark                  old allocs     new allocs     delta
BenchmarkReadFilter-12     1350688        833994         -38.25%

benchmark                  old bytes      new bytes     delta
BenchmarkReadFilter-12     1379473016     577756776     -58.12%
```

### Switch the order of t0 and t1 so t0 has a larger cardinality

```
jsternberg:[~/g/s/g/i/influxdb] $ (perf/table-read *$<>) benchcmp old.txt new3.txt
benchmark                  old ns/op      new ns/op     delta
BenchmarkReadFilter-12     1291650820     343158702     -73.43%

benchmark                  old allocs     new allocs     delta
BenchmarkReadFilter-12     1350688        746656         -44.72%

benchmark                  old bytes      new bytes     delta
BenchmarkReadFilter-12     1379473016     420456485     -69.52%
```

This optimization is heavily dependent on the order of the tables coming out of storage, but it does seem like it improves all situations. In the future, we can also improve the performance here by lazily creating the arrow arrays as shown [here](https://github.com/influxdata/flux/pull/2190). That should reduce the need for the lru cache, but not eliminate the need because we would still want to share the column data between these for when the actual column is materialized.