@jsternberg 2019-11-25
Also to clarify, there's no assumptions about the order of how stuff comes out of the cache. I was just noting that the performance can change depending on that order so it's a bit hard to benchmark. I added that note as a comment.